Move path-related logic into the path directory
