+++ packages/arbor-store/src/index.ts
@@ -6,7 +6,7 @@ export { ArrayNodeHandler } from "./handlers/ArrayNodeHandler"
 export { MapNodeHandler } from "./handlers/MapNodeHandler"
 export { NodeHandler } from "./handlers/NodeHandler"
 export * from "./MutationEngine"
-export { Path } from "./Path"
+export { Path } from "./path"
 export * from "./ScopedStore"
 export { Subscriptions } from "./Subscriptions"
 export type * from "./types"
