+++ packages/arbor-store/tests/Subscribers.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it, vi } from "vitest"
 
 import { Arbor } from "../src/Arbor"
-import { Path } from "../src/Path"
+import { Path } from "../src/path/Path"
 import { Subscriptions } from "../src/Subscriptions"
 
 describe("Subscriptions", () => {
